Update Get-LastSystemEvents.psm1
With this version, the user can choose which event log they want to view and how many events they want to retrieve. The -LogName parameter is for specifying the event log name and -MaxEvents parameter for the number of events. The default values are "System" and 10, respectively, if the user doesn't specify them.